refactor(ipc): extract meeting commands into commands/meeting.rs (#82)
The IPC layer's `commands.rs` has grown from 1378 → 2085 lines
since #82 was filed (+51% LOC, +48% commands). Meeting Mode is the
biggest cohesive group inside it — 7 `#[tauri::command]` functions,
3 result types, the `MAX_MEETING_SOURCES` cap + `sanitise_*`
validator, plus six tests for the validator. All of it has clear
domain seam (touches `state.meeting_manager`, `state.meetings`,
`crate::hud`, the `MeetingSourceFailedPayload` event channel) and
nothing in the rest of the file imports from it.

Move:

- `commands.rs` → `commands/mod.rs` (file rename only).
- Meeting Mode block → new `commands/meeting.rs`. `MeetingSessionDetail`,
  `ActiveMeetingSession`, `MAX_MEETING_SOURCES`, `sanitise_meeting_sources`,
  and the seven commands move with their docstrings intact.
- The `sanitise_*` tests move alongside the validator under
  `meeting.rs`'s own `mod tests` block.
- `lib.rs` `generate_handler!` invocations updated from
  `ipc::commands::meeting_*` to `ipc::commands::meeting::meeting_*`.
  Per `learnings.md` (2026-04-25), the macro is path-sensitive and
  does NOT see commands through `pub use` re-exports — the hidden
  `__cmd__<name>` symbol stays in the defining module. Module
  header comment in `commands/mod.rs` cites this so a future
  contributor doesn't try the re-export path.

This is the focused-extraction shape #82's body explicitly preferred
("don't preemptively split for stylistic reasons" — but the next
domain seam, when it appears, is the trigger to act). Future
extractions of Models / History / Replacements / Vocabulary can
follow the same pattern PR-by-PR if friction grows.

### Smoke

- `cargo build`, `cargo clippy --all-targets -- -D warnings` clean
- `cargo test --lib` 206 passed (the 6 sanitise_* tests now live
  in `meeting::tests`, plus the rest of the suite unchanged)
- `npm run tauri dev` launches; meeting panel + start/stop work.
