+++ src-tauri/src/lib.rs
@@ -188,13 +188,13 @@ pub fn run() {
             ipc::commands::open_macos_privacy_pane,
             ipc::commands::diagnose_macos_permissions,
             ipc::commands::reset_macos_permissions,
-            ipc::commands::meeting_sessions_list,
-            ipc::commands::meeting_session_get,
-            ipc::commands::meeting_session_delete,
-            ipc::commands::meeting_session_set_notes,
-            ipc::commands::meeting_active_session,
-            ipc::commands::meeting_start_manual,
-            ipc::commands::meeting_stop_manual,
+            ipc::commands::meeting::meeting_sessions_list,
+            ipc::commands::meeting::meeting_session_get,
+            ipc::commands::meeting::meeting_session_delete,
+            ipc::commands::meeting::meeting_session_set_notes,
+            ipc::commands::meeting::meeting_active_session,
+            ipc::commands::meeting::meeting_start_manual,
+            ipc::commands::meeting::meeting_stop_manual,
         ])
         .run(tauri::generate_context!())
         .expect("error while running Hush");
